criando services, métodos de mapping em entidades e corrigindo tipo de id dos dtos

+++ src/main/java/br/edu/ifs/apinewsigaa/rest/dto/ProfessorDto.java
@@ -14,8 +14,7 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProfessorDto {
-
-    private long id;
+    private int id;
 
     @NotBlank(message = "Nome é obrigatório")
     private String nome;
